Prevents collisions between native HTML5 input range elements and `ampStack` gesture detection
To recreate issue:

- create an `ampStack` with gesture navigation enabled
- introduce an element `<input type="range" class="amp-spin-slider" value="1" min="1" max="50" step="1" />` and position it using `position: absolute` over a stack element.
- try to move the slider on a non-touch device, the slider works...
- try to move it using a touch device, the stack will incorrectly invoke `goTo` and navigate based on the slide gesture because it is picking up the event.

Please note:  The slider (input type=range) HTML5 element relies on the event being propogated, so cancelling the events using a container so that they don't reach the stack isn't a viable option, the slider will simply not work by doing that.

The approach taken instead is to ensure that gestures are not coming from an original target of an html5 slider element.