ensure that `_onImageLoaded` is only actioned once
`_onImageLoaded` is bound to the `load` event of the target image to use for zoom, and the Image's load event seems to fire each time the user switches between alternate images in a stack (see https://ampliencedev.atlassian.net/browse/SDK-165 for example case of this bug).

This results in events being bound multiple times on the same element and race conditions can occur where the zoom is toggled off before it gets a chance to perform its zooming `setInterval` routine.

This fix ensures that if the image is loaded, it will not rebind events.