applies width and height modifiers *after* template specification on the querystring
Since DM disregards `w` and `h` modifiers if they are in the querystring before template specification, inline zoom on templated images would serve the full size image scaled client side to the given width and height.

What we wanted was for the template to be applied and for DM to also apply the width and height values to ensure that the image returned from the dynamic imaging server was of the correct size.

This fix ensures that parameter order is presented in a way for DM to do this.