allows for optionally disabling soft states for the amp video widget
`ampVideo` allows for "soft states" to be ignored and not tracked, for example, when a video is set to loop it will not raise another play event on the second loop.

However, depending on your usage of `ampVideo` it is possible to get the viewer into a condition where states are not being tracked (as they are interpretted as soft) and so the player doesn't reflect the state you anticipate.  By disabling soft states, you will receive more notifications, but also always get the correct state information.

This has been made an option so as not to affect any viewers that already use the widget in the current way (ie. with soft states)