adds `currentZoomState` to inline zoom widget to return the zoom state
tracking provides an easy way to retrieve zoom state after an event happens, but to store this state on the client creates a state duplication and can lead clients to potential state corruption / getting out of sync.  By creating a method which can be interrogated by the client to get the current state, we can mitigate this.