boundary margins subtract
Determination of visible assets is not correct, because:

elmSize - is going with it's paddings and margins, so if we have 3 assets in widget , it consist from - 3 base width of asset and stacked margin and padding. We should consider internal inter-element margins and paddings, but the wrapped widget ignores left and right boundary margins. As a result some assets are going to partially visible even if they have to be visible. 

The solution is to subtract boundary margins in all operations related to calculation visible elements.